+++ tests/PreparedQueryTest.php
@@ -4,23 +4,11 @@ namespace YdbPlatform\Ydb\Test;
 
 use PHPUnit\Framework\TestCase;
 use YdbPlatform\Ydb\Auth\Implement\AnonymousAuthentication;
-use YdbPlatform\Ydb\Table;
 use YdbPlatform\Ydb\Ydb;
 
-class SessionManager extends \YdbPlatform\Ydb\Session{
-    public static function setSessionId(\YdbPlatform\Ydb\Session $session, string $id){
-        $session->session_id = $id;
-        return $session;
-    }
-    public static function getSessionId(\YdbPlatform\Ydb\Session $session){
-        return $session->session_id;
-    }
-}
-
-class RetryOnBadSessionTest extends TestCase
+class PreparedQueryTest extends TestCase
 {
     public function test(){
-
         $config = [
 
             // Database path
@@ -40,15 +28,20 @@ class RetryOnBadSessionTest extends TestCase
         ];
         $ydb = new Ydb($config);
         $table = $ydb->table();
+
         $session = $table->createSession();
-        $oldSessionId = SessionManager::getSessionId($session);
-        $session->delete();
-        $session = $table->createSession();
-        SessionManager::setSessionId($session,$oldSessionId);
-        $tres = $session->query('select 1 as res')->rows()[0]['res'];
-        self::assertEquals(
-            1,
-            $tres
+
+        $prepared_query = $session->prepare('
+declare $pk as Int64;
+select $pk;');
+        $x = 2;
+        $result = $session->transaction(function($session) use ($prepared_query, $x){
+            return $prepared_query->execute([
+                'pk' => $x,
+            ]);
+        });
+        self::assertEquals($x,
+            $result->rows()[0]['column0']
         );
     }
 }